moving texts to an typography package
