+++ app/src/main/java/com/marelso/designsystem/ui/home/HomeScreenHoisting.kt
@@ -18,5 +18,6 @@ fun HomeScreenHoisting() {
         HomeSubtitleComponent(modifier = Modifier.padding(bottom = dimen2))
         HomeBodyComponent(modifier = Modifier.padding(bottom = dimen4))
         HomeTitleComponent(modifier = Modifier.padding(bottom = dimen2))
+
     }
 }
