+++ app/src/main/java/com/marelso/designsystem/ui/home/components/HomeBodyComponent.kt
@@ -2,7 +2,7 @@ package com.marelso.designsystem.ui.home.components
 
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import com.marelso.designsystem.components.BodyComponent
+import com.marelso.designsystem.components.typography.BodyComponent
 
 @Composable
 fun HomeBodyComponent(modifier: Modifier = Modifier) {
