+++ app/src/main/java/com/marelso/designsystem/ui/home/components/HomeHeadlineComponent.kt
@@ -2,7 +2,7 @@ package com.marelso.designsystem.ui.home.components
 
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import com.marelso.designsystem.components.HeadlineComponent
+import com.marelso.designsystem.components.typography.HeadlineComponent
 
 @Composable
 fun HomeHeadlineComponent(modifier: Modifier = Modifier) {
