+++ app/src/main/java/com/marelso/designsystem/ui/home/components/HomeSubtitleComponent.kt
@@ -2,7 +2,7 @@ package com.marelso.designsystem.ui.home.components
 
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import com.marelso.designsystem.components.SubtitleComponent
+import com.marelso.designsystem.components.typography.SubtitleComponent
 
 @Composable
 fun HomeSubtitleComponent(modifier: Modifier = Modifier) {
