+++ app/src/main/java/com/marelso/designsystem/ui/home/components/HomeTitleComponent.kt
@@ -2,7 +2,7 @@ package com.marelso.designsystem.ui.home.components
 
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
-import com.marelso.designsystem.components.TitleComponent
+import com.marelso.designsystem.components.typography.TitleComponent
 
 @Composable
 fun HomeTitleComponent(
